Extract agent task recipe builder
