+++ packages/runtime-core/package.json
@@ -12,6 +12,10 @@
     "./recipe-builders": {
       "types": "./dist/recipe-builders.d.ts",
       "import": "./dist/recipe-builders.js"
+    },
+    "./agent-task-recipe": {
+      "types": "./dist/agent-task-recipe.d.ts",
+      "import": "./dist/agent-task-recipe.js"
     }
   },
   "files": [
