+++ packages/runtime-core/src/index.ts
@@ -14,6 +14,7 @@ export * from "./task-input.js"
 export * from "./structured-artifacts.js"
 export * from "./agent-task-run-result.js"
 export * from "./agent-runtime-workload.js"
+export * from "./agent-task-recipe.js"
 export * from "./browser-interaction.js"
 export * from "./browser-review-bridge.js"
 export * from "./browser-result-shapes.js"
